move new TaggedItemCounts to its own file
